TST: Skip test_suspenders if pcaspy is not available
