+++ bluesky/tests/test_suspenders.py
@@ -3,6 +3,8 @@ import sys
 import asyncio
 import time as ttime
 import pytest
+from multiprocessing import Process
+import signal
 from bluesky import Msg
 from bluesky.tests.utils import setup_test_run_engine
 
@@ -10,6 +12,45 @@ from bluesky.tests.utils import setup_test_run_engine
 RE = setup_test_run_engine()
 loop = asyncio.get_event_loop()
 
+pcaspy_process = None
+
+def setup_module():
+    try:
+        from pcaspy import Driver, SimpleServer
+    except ImportError as ie:
+        pytest.skip("pcaspy is not available. Skipping all suspenders test."
+                    "ImportError: {}".format(ie))
+        def to_subproc():
+
+            prefix = 'BSTEST:'
+            pvdb = {
+                'VAL': {
+                    'prec': 3,
+                },
+            }
+
+            class myDriver(Driver):
+                def __init__(self):
+                    super(myDriver, self).__init__()
+
+            server = SimpleServer()
+            server.createPV(prefix, pvdb)
+            driver = myDriver()
+
+            # process CA transactions
+            while True:
+                try:
+                    server.process(0.1)
+                except KeyboardInterrupt:
+                    break
+
+        pcaspy_process = Process(target=to_subproc)
+        pcaspy_process.start()
+
+def teardown_module():
+    if pcaspy_process is not None:
+        os.kill(ppcaspy_processpid, signal.SIGINT)
+        pcaspy_process.join()
 
 def test_epics_smoke():
